Register independent vars on each install type
When jenkins install is called, both tasks are visited. And
it registers result of jenkins_install_package twice, causing
it to have a wrong value, and don't install jenkins users, causing
security to don't be configure and plugins to fail.

It's better to add different vars for each task, and trigger
the users install in case of some of the two changed.