Home directory support
Introduced the jenkins_home variable which follows the standard
behaviour of the JENKINS_HOME environment variable. This allows users
to specify a different volume to store build artefacts, logs and
archives.

The default value for the jenkins_home variable is ‘/var/lib/jenkins”
meaning that the default behaviour is respected.