Allow plugin install to work using pinned versions
When we added the ability to pin versions, it broke the ability to install
plugins since it doesn't register the variables correctly to trigger the
changes introduced in #72.

Unfortunately, the fix of adding a register to the pinned version package
installation module (yum / apt) execution re-introduces the exact same issue
that #72 is attempting to resolve.

So, in order to work around that limitation, I've moved the deployment of
the authentication groovy script to a handler, and then notified that handler
from both the application installation routines.

There is also the addition of a meta module to flush the handlers so that
the groovy script is installed immediately, and thereby registering the
correct values to trigger a restart of Jenkins if necessary.

Resolves #70