Update Travis-CI to execute available tests
Update the Travis-CI file so that all available tests are
now setup to run. Fixes a bug with one of the tests that
wasn't properly configured to run the plugins test.

Also cleaned up the formatting of the file so that it was more
obvious what blocks are deploying which tests. Made sure that
all tests were executed on both Ubuntu and CentOS systems.