Add support for setting a public SSH key for the admin user
According to https://wiki.jenkins-ci.org/display/JENKINS/Jenkins+CLI
"if your Jenkins requires authentication, you should set up public
key authentication". This is particularly true for certain aspects
of Jenkins CLI, like creating a job, which will awkwardly fail.

Adding this option helps with further automation of Jenkins.
It should be kept in mind that a restart seems to be required for
changes to apply.