Jenkins needs to be restarted after installation of plugins
As you might now, handlers runs at the very end of the playbook.
That causes issues while you are using this role and trying to configure
Jenkins with the next role/tasks.

Jenkins restart might happen just after the installation of the plugins.
So that, I can configure some of the plugins by jenkins-cli