#304: Configure Jenkins' updates URL
Some corporate firewalls restrict outbound traffic.  Jenkins default uses a
mirror system where the list of possible endpoints that updates and plugins may
be downloaded from is changeable with little or no notice.  Maintaining the
firewall's permitted URL list when the mirror list changes may therefore be
problematic, requiring change tickets etc.

Whilst we did have the ability to download the updates list from the specified
mirror, when jenkins was asked to install a given plugin, the list still
referred to updates.jenkins-ci.org, which might then fail.  This change permits
a modified version of the updates file to be used with a fixed mirror in it to
safely negotiate such a corporate firewall.