Adds the ability to enable proxy compatibility
In some cases it is not possible to provide proxy configuration because
it is a 3rd party that is out of your control or you are using some port
forwarding service that has a block of IPs (multiple proxy IPs). Jenkins
provides a proxy compatibility mode that excludes the client IP when
creating the CSRF token. This option needs to be exposed via the role
as it cannot be set via the web browser after completing setup because
any attempt to change the config will faile the CSRF check :)