tasks: Do not unnecessarily restart Jenkins twice
When introducing support for a user-specified http_port, ef68aa8c
introduced an early restart of Jenkins in order to differentiate between
a jenkins_init_config change and a jenkins_http_config change. Later on,
in commit e4cd4b9c, a more interesting logic appeared to do the restart
whether one or another thing has happened.

Since then, separate restarts are not really needed anymore. This
removes the first one and moves the logic down to the end of the tasks.

Signed-off-by: Felipe Franciosi <felipe@nutanix.com>