Add HTTPS support using OpenSSL cert/key pairs
Using OpenSSL is preferred instead of a Java keystore (JKS) because JKSs
require a passphrase and this role does not use an argument file. As a
result, the passphrase would have to be set in the init file and end up
exposed in the Jenkins command line.

To use the feature, a user must produce a ca-signed .crt file and a .key
file which is vault-encrypted. Their local paths are set via:
    jenkins_https_crt
    jenkins_https_key

The user must also specify a port which enables the feature. Note that
the key is kept unencrypted at the Jenkins controller, but protected so
it is only accessible via the Jenkins user. This is also common practice
with other HTTPS servers. The port is specified via:
    jenkins_https_port

When using -1, the support is disabled.

Signed-off-by: Felipe Franciosi <felipe@nutanix.com>