Jenkins dropped support for --httpsPrivateKey and --httpsCertificate
As of Jenkins 2.363, we need to add certificates to a keystore
and pass --httpsKeyStore=path/to/keystore and
--httpsKeyStorePassword=keystorePassword instead.

This change uses the jenkins_https_crt and jenkins_https_key params
to generate a new keystore, with jenkins_keystore_pass as the
password to encrypt it. If jenkins_https_port is set to '-1',
these arguments are not are not passed to Jenkins at startup.

Note that the keystore password will be stored in plaintext on disk
and will appear in the process arguments.

Signed-off-by: Tim Foster <tim.foster@nutanix.com>